+++ src/Frontend/ValueFilter.php
@@ -73,6 +73,8 @@ class ValueFilter
      */
     public function filter($value)
     {
-        return $this->getReplacer()->replace($value);
+        $value = $this->getReplacer()->replace($value, false);
+
+        return $value;
     }
 }