+++ src/Mapper/Layer/MarkersLayerMapper.php
@@ -13,7 +13,6 @@
 namespace Netzmacht\Contao\Leaflet\Mapper\Layer;
 
 use Contao\Model;
-use Netzmacht\Contao\Leaflet\Frontend\RequestUrl;
 use Netzmacht\Contao\Leaflet\Mapper\DefinitionMapper;
 use Netzmacht\Contao\Leaflet\Mapper\GeoJsonMapper;
 use Netzmacht\Contao\Leaflet\Mapper\Request;
@@ -24,6 +23,7 @@ use Netzmacht\LeafletPHP\Definition;
 use Netzmacht\LeafletPHP\Definition\Group\GeoJson;
 use Netzmacht\LeafletPHP\Plugins\Omnivore\GeoJson as OmnivoreGeoJson;
 use Netzmacht\LeafletPHP\Value\GeoJson\FeatureCollection;
+use Symfony\Component\Routing\RouterInterface;
 
 /**
  * Class MarkersLayerMapper maps the layer model to the markers definition.
@@ -47,13 +47,22 @@ class MarkersLayerMapper extends AbstractLayerMapper implements GeoJsonMapper
     private $repositoryManager;
 
     /**
+     * Router.
+     *
+     * @var RouterInterface
+     */
+    private $router;
+
+    /**
      * Construct.
      *
      * @param RepositoryManager $repositoryManager Repository manager.
+     * @param RouterInterface   $router            Router.
      */
-    public function __construct(RepositoryManager $repositoryManager)
+    public function __construct(RepositoryManager $repositoryManager, RouterInterface $router)
     {
         $this->repositoryManager = $repositoryManager;
+        $this->router            = $router;
 
         parent::__construct();
     }
@@ -93,7 +102,7 @@ class MarkersLayerMapper extends AbstractLayerMapper implements GeoJsonMapper
 
                 return [
                     $this->getElementId($model, $elementId),
-                    RequestUrl::create($model->id, null, null, $request),
+                    $this->generateUrl((int) $model->id, $request),
                     [],
                     $layer,
                 ];
@@ -101,7 +110,7 @@ class MarkersLayerMapper extends AbstractLayerMapper implements GeoJsonMapper
 
             return [
                 $this->getElementId($model, $elementId),
-                RequestUrl::create($model->id, null, null, $request),
+                $this->generateUrl((int) $model->id, $request),
             ];
         }
 
@@ -176,10 +185,37 @@ class MarkersLayerMapper extends AbstractLayerMapper implements GeoJsonMapper
     {
         $repository = $this->repositoryManager->getRepository(MarkerModel::class);
 
-        if ($model->boundsMode == 'fit') {
+        if ($model->boundsMode === 'fit') {
             return $repository->findByFilter($model->id, $request->getFilter());
         }
 
         return $repository->findByFilter($model->id);
+    }
+
+    /**
+     * Generate the data url for a layer.
+     *
+     * @param int          $layerId The layer id.
+     * @param Request|null $request The request.
+     *
+     * @return string
+     *
+     * @SuppressWarnings(PHPMD.Superglobals)
+     */
+    private function generateUrl(int $layerId, ?Request $request = null): string
+    {
+        $params = ['layerId' => $layerId];
+
+        if ($request && ($filter = $request->getFilter())) {
+            $params['filter'] = $filter::getName();
+            $params['values'] = $filter->toRequest();
+        }
+
+        if (isset($GLOBALS['objPage'])) {
+            $params['context']   = 'page';
+            $params['contextId'] = $GLOBALS['objPage']->id;
+        }
+
+        return $this->router->generate('leaflet_layer', $params, RouterInterface::ABSOLUTE_URL);
     }
 }